+++ libs/data-mapper-v2/src/components/common/reactflow/SchemaNode.tsx
@@ -1,15 +1,18 @@
-import { Handle, Position, useUpdateNodeInternals, type NodeProps } from 'reactflow';
+import { Handle, Position, useEdges, useUpdateNodeInternals, type NodeProps } from 'reactflow';
 import type { SchemaNodeReactFlowDataProps } from '../../../models/ReactFlow';
 import { mergeClasses } from '@fluentui/react-components';
 import { useStyles } from './styles';
-import { useRef, useEffect } from 'react';
+import { useRef, useEffect, useMemo } from 'react';
 
 const SchemaNode = (props: NodeProps<SchemaNodeReactFlowDataProps>) => {
   const divRef = useRef<HTMLDivElement | null>(null);
-  const updateNodeInternals = useUpdateNodeInternals();
   const { data, id } = props;
-  const { isLeftDirection, isConnected } = data;
+  const { isLeftDirection } = data;
+  const updateNodeInternals = useUpdateNodeInternals();
+  const edges = useEdges();
   const styles = useStyles();
+  // danielle update this to move away from edges
+  const isConnected = useMemo(() => edges.some((edge) => edge.source === id || edge.target === id), [edges, id]);
 
   useEffect(() => {
     updateNodeInternals(id);
@@ -20,9 +23,7 @@ const SchemaNode = (props: NodeProps<SchemaNodeReactFlowDataProps>) => {
         type={isLeftDirection ? 'source' : 'target'}
         position={Position.Left}
         className={mergeClasses(styles.handleWrapper, isConnected ? styles.handleConnected : '')}
-        isConnectableStart={isLeftDirection}
-        isConnectableEnd={!isLeftDirection}
-        style={{ left: '-7px' }}
+        isConnectable={true}
       />
     </div>
   );