+++ libs/designer-ui/src/lib/nodeCollapseToggle/index.tsx
@@ -1,12 +1,13 @@
 import { TooltipHost, DirectionalHint, Icon, css } from '@fluentui/react';
 import { FontSizes } from '@fluentui/theme';
+import { useCardKeyboardInteraction } from '../card/hooks';
 import { useIntl } from 'react-intl';
 
 interface NodeCollapseToggleProps {
   disabled?: boolean;
   collapsed?: boolean;
   onSmallCard?: boolean;
-  handleCollapse?: (event: { currentTarget: any }) => void;
+  handleCollapse?: () => void;
 }
 
 const NodeCollapseToggle = (props: NodeCollapseToggleProps) => {
@@ -28,6 +29,8 @@ const NodeCollapseToggle = (props: NodeCollapseToggleProps) => {
   const iconName = collapsed ? 'ChevronDown' : 'ChevronUp';
   const toggleText = collapsed ? EXPAND_TEXT : COLLAPSE_TEXT;
 
+  const keyboardInteraction = useCardKeyboardInteraction(handleCollapse);
+
   return (
     <TooltipHost content={toggleText} directionalHint={DirectionalHint.rightCenter}>
       <button
@@ -35,6 +38,8 @@ const NodeCollapseToggle = (props: NodeCollapseToggleProps) => {
         disabled={disabled}
         className={css('msla-collapse-toggle', disabled && 'disabled', onSmallCard && 'small')}
         onClick={handleCollapse}
+        onKeyDown={keyboardInteraction.keyDown}
+        onKeyUp={keyboardInteraction.keyUp}
       >
         <Icon iconName={iconName} styles={{ root: { fontSize: FontSizes.small } }} />
       </button>